Ignore possibility for Javascript serialization
Added an annotation that could be used to tag elements (fields and methods) which should be ignored when serializing them to Javascript. This can be espacially useful for easy performance enhancements or when lazy loading objects with circulair references, which aren't necessary. In my case lazy loading objects sometimes caused the Javascript serializer to try and serialize 'everything' in memory, which caused a lock up.

I've currently not added a test or test-cases and thus am not sure if these proposed changes are 100% correct. These should probably be tested, but this should be an easy extra and start.