Fix logging output in AuthUtils
AuthUtils.getAuthenticationObject() has a logging statement that ends up calling:

logger.trace(msg, new Object[] {obj1, obj2}, obj3)

But, that doesn't work like the author intended -- it treats the Object[] as one item to log, and obj3 as another. so the first placeholder gets replaced with [obj1, obj2], the second with obj3, and the 3rd placeholder just stays as "{}".