First iteration of Spring 5 support
This is basically a copy of the Spring 4 support with modified
dependencies and package names.

The DOCTYPE translator processor has been removed
because it was meant to work as a legacy integration with
Thymeleaf 2.1 DTDs. But the Spring 5 module will not be
compatible with Thymeleaf 2.1 (and anyway, if anyone needed
to use Thymeleaf 2.1 with Spring 5 they could simply use
the thymeleaf-spring4 packages, which would work OK and
should only make them lose reactive support).

NOTE this is depending on Spring 5.0's snapshots, not the
current milestone M3. This is due to things in the current M3
that would make the integration of the reactive parts
impossible (and that will be fixed in M4).

NOTE this still does not contain the reactive integrations
developed separately at the thymeleafsandbox-springreactive
repository. They will be included afterwards here, but
importantly the spring-web-reactive dependency should be
kept OPTIONAL.