Steps towards processing SSE output at the highest-level ThrottledTemplateWriter level
The intention of these modifications is to pave the way to allow several SSE events to be rendered in a
single buffer for better performance in non-tailed-cursor scenarios. This should allow direct interaction
between the DataDrivenIterator and the SSEThrottledTemplateWriter so that the iterator, when signaled
when an iteration has really finished (by the gathered model), can in turn ask the writer to finish by
writing its overflow buffer and then closing the SSE event.